Disables track resizing in leaflet.
When the map is resized, a `resize` and a `movend` event are raised. Several features
on the app are based on `movend` events, so this "after-resize" behaviour is dangerous.

Mainly because in some browsers downloading a document shows a status bar at the
bottom of the browser that can forzes the map to be resized. And undesired behaviour
arises in the app.