Bugfix. Adds  ON UPDATE CASCADE to new 'unidade' field on user table
Commonly this is the desired behaviour. If a value in the 'unidade' domain must
be changed, we will want that this change will be reflected on users table.

Also, in the case of a DELETE, we want it don't allow the change if there are
depending users to obligate the developer to handle it properly

refs #1582