Defined column models will eagerly get their default value
Having strict constraints in the database (like NOT NULL) is a good thing. But,
database defaults are not usually enought to avoid exceptions. Model instances
are generated in the backend without defaults, send to the client, and received
again.

As soon as posible we set default valid values the better.

refs #3159