Renovacoes api now returns the full Exploracao
PUT, PATCH requests to /api/renovacoes return only the Renovacao data and not
the full representation of the Exploracao. This is not consistent with GET
that returns the full representation.

Also, there is a bug on this, geometry is not returned and after the save
and the requests going through all `parse` methods, the original geometry
in the model is overwritten with an empty generic Backbone.Model.

refs #3492